+++ Assets/Script/Monster/Model/IMonster.cs
@@ -6,6 +6,6 @@ public interface IMonster
 {
         void Move();
         void SpecialAbility();
-        void TakeDamage(int damage);
+        void TakeDamage(float damage);
         void Die();
 }
